refactor(kernel): split policy tests
Move the large inline PolicyPipeline test module out of policy.rs into policy/tests.rs.

Keep the change mechanical: production policy code is unchanged, and the parent module now only declares the test module.
